requirements.txt
pyTelegramBotAPI
python-dotenv
APScheduler
requests
tzdata
